Convert forgotten logf to log.

+++ src/rest_server/rest_server.cpp
@@ -435,7 +435,7 @@ BOOL WINAPI RestServerCtrlCHandler(DWORD dwCtrlType) {
 }
 
 bool rest_server::wait_until_signalled() {
-  logf("Waiting until Ctrl+C is pressed.");
+  log("Waiting until Ctrl+C is pressed.");
   unique_lock<mutex> lock(rest_server_ctrl_c_mutex);
   if (!rest_server_ctrl_c_handlers) {
     rest_server_ctrl_c_signalled = false;
